associate user with task

+++ src/tasks/tasks.service.ts
@@ -5,7 +5,8 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
-import { stat } from 'fs';
+import { User } from 'src/auth/user.entity';
+import { userInfo } from 'os';
 
 @Injectable()
 export class TasksService {
@@ -14,50 +15,31 @@ export class TasksService {
     private taskRepository: TaskRepository
   ) { }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id)
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } })
     if (!found) {
       throw new NotFoundException(`Task with id ${id} not found`)
     }
     return found
   }
 
-  getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto)
+  getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user)
   }
 
-  /* getAllTasks(): Task[] {
-    return this.tasks;
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user)
   }
 
-  getTasksWithFilters(filterDto: GetTasksFilterDto): Task[] {
-    const { status, search } = filterDto
-    let tasks = this.getAllTasks()
-
-    if (status) {
-      tasks = tasks.filter(task => task.status === status)
-    }
-
-    if (search) {
-      tasks = tasks.filter(task => task.title.includes(search) || task.description.includes(search))
-    }
-
-    return tasks
-  }*/
-
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto)
-  }
-
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id)
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id })
     if (result.affected === 0) {
       throw new NotFoundException(`Task with id ${id} not found`)
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id)
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user)
     return this.taskRepository.updateTaskStatus(task, status)
   }
 }
